% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_medicare_stats.R
\name{get_medicare_stats}
\alias{get_medicare_stats}
\title{Calculate Medicare Statistics}
\usage{
get_medicare_stats(df, stat = "mean")
}
\arguments{
\item{df}{dataframe that contains the following columns: 'DRG.Definition',
'Average.Covered.Charges', 'Average.Total.Payments', and
'Average.Medicare.Payments'.}

\item{stat}{string specifying statistic to calculate. Accepts 'mean',
'median', or 'sd'. Default: 'mean'.}
}
\value{
double of specified statistic (mean, median, or sd)
}
\description{
This function calculates statistics over all the DRG codes for average
Medicare payments. It can calculate either the mean, median, or standard
deviation of the DRG codes.
}
\examples{
df.test <- drg
get_medicare_stats(drg, stat = "sd")
}
