% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_payments.R
\name{plot_payments}
\alias{plot_payments}
\title{Box Plot of Average Payments}
\usage{
plot_payments(df, pay_type = "medicare")
}
\arguments{
\item{df}{dataframe that contains the following columns: 'DRG.Definition',
'Average.Covered.Charges', 'Average.Total.Payments', and
'Average.Medicare.Payments'.}

\item{pay_type}{string of payment type. Accepts 'covered' for average
covered charges, 'total' for average total payment, and 'medicare' for
average Medicare payments. Default: 'medicare'.}
}
\value{
boxplot of average payments for all DRG codes
}
\description{
This function produces a boxplot of payments by DRG code. It takes in an
argument to calculate the payments for either the average covered charges,
the average total payment, or the average Medicare payments.
}
\examples{
df.test <- drg
plot_payments(df.test, pay_type = "covered")

}
